Removed logica from UserController to UserService Extended POST with
Validation on inputDTO
Added Exeption handeling for user must be Unique
